2024_07_11/requirements.txt
requests
pydantic
psycopg2-binary
python-dotenv
flask
gunicorn